refactor: 해더, 푸터, 인트로 파일명 변경 및 루트 수정
