+++ project05/app/layout.js
@@ -1,7 +1,7 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Header from "@/src/components/header";
-import Footer from "@/src/components/footer";
+import Header from "@/src/components/Header";
+import Footer from "@/src/components/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
