+++ project05/app/page.js
@@ -1,10 +1,10 @@
 import Core from "@/src/components/sections/Core";
-import MainIntro from "@/src/components/sections/MainIntro";
+import Intro from "@/src/components/sections/Intro";
 
 export default function Home() {
   return (
     <div className="h-[2000px] bg-[#121212]">
-      <MainIntro />
+      <Intro />
       <Core />
     </div>
   );
